requirements.txt
fire==0.6.0
gradio==4.39.0
loguru==0.7.2
Pillow==10.4.0
setuptools==69.5.1
tiktoken==0.5.2
python-dotenv==0.13.0
openai==1.30.1
WorkWeixinRobot
bs4
schedule
tinydb
daily-basic-function @ git+https://github.com/aigc-open/daily-basic-function.git
global-config @ git+https://github.com/aigc-open/global-config.git
duckduckgo_search
serpapi==0.1.5
wechatpy
pydantic
requests
